--- applications/services/cli/cli_command_sysctl.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include "cli_i.h"
-
-void cli_command_sysctl(Cli* cli, FuriString* args, void* context);
