+++ applications/services/cli/cli_commands.c
@@ -5,7 +5,7 @@
 #include "cli_command_light_sensor.h"
 #include "cli_command_audio.h"
 #include "cli_command_sl_cli.h"
-#include "cli_command_sysctl.h"
+#include "cli_command_factory_reset.h"
 
 #include <core/thread.h>
 #include <core/thread_list.h>
@@ -54,6 +54,52 @@ void cli_command_help(Cli* cli, FuriString* args, void* context) {
         printf("%s", furi_string_get_cstr(args));
         printf("` command not found");
     }
+}
+
+static void cli_command_sysctl_debug(Cli* cli, FuriString* args, void* context) {
+    UNUSED(context);
+
+    if(furi_string_equal_str(args, "0")) {
+        cli_delete_command(cli, "gpio");
+        cli_delete_command(cli, "sl_echo");
+        cli_delete_command(cli, "factory_reset");
+        printf("Debug disabled.");
+    } else if(furi_string_equal_str(args, "1")) {
+        cli_add_command(cli, "gpio", CliCommandFlagParallelSafe, cli_command_gpio, NULL);
+        cli_add_command(cli, "sl_echo", CliCommandFlagParallelSafe, cli_command_sl_echo, NULL);
+        cli_add_command(
+            cli, "factory_reset", CliCommandFlagParallelSafe, cli_command_factroy_reset, NULL);
+        printf("Debug enabled.");
+    } else {
+        cli_print_usage("sysctl debug", "<1|0>", furi_string_get_cstr(args));
+    }
+}
+
+static void cli_command_sysctl_print_usage() {
+    printf("Usage:\r\n");
+    printf("sysctl <cmd>\r\n");
+    printf("Cmd list:\r\n");
+    printf("\tdebug - enables or disables some debug commands\r\n");
+}
+
+void cli_command_sysctl(Cli* cli, FuriString* args, void* context) {
+    FuriString* cmd;
+    cmd = furi_string_alloc();
+
+    do {
+        if(!args_read_string_and_trim(args, cmd)) {
+            cli_command_sysctl_print_usage();
+            break;
+        }
+
+        if(furi_string_cmp_str(cmd, "debug") == 0) {
+            cli_command_sysctl_debug(cli, args, context);
+            break;
+        }
+        cli_command_sysctl_print_usage();
+    } while(false);
+
+    furi_string_free(cmd);
 }
 
 void cli_command_uptime(Cli* cli, FuriString* args, void* context) {
